requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
openai==1.61.0
python-dotenv==1.0.1
mistralai
piper-tts==1.2.0
